Moved block and version messages and shared utils into seperate files.
